Feature/moving core to mp kotlin

+++ core-coinbase-pro/src/main/kotlin/com/dllewellyn/coinbaseapi/adapter/AccountsAdapter.kt
@@ -20,5 +20,5 @@ class AccountsAdapter(
             } ?: account.toCore()
         }
 
-    override suspend fun getNonEmptyAccounts() = getAccounts().filter { a -> a.balance.toInt() > 0 }
+    override suspend fun getNonEmptyAccounts() = getAccounts().filter { a -> a.balance > 0 }
 }
